project_module_4/app/requirements.txt
SQLAlchemy==2.0.41
transformers==4.53.2
h5py==3.14.0
wrapt==1.17.2
protobuf==6.31.1
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
google==3.0.0
bcrypt==4.3.0
pydantic-settings==2.10.1
sqlmodel==0.0.24
loguru==0.6.0
psycopg2-binary==2.9.10
fastapi[standard-no-fastapi-cloud-cli]==0.116.1
uvicorn==0.35.0
pyjwt==2.10.1
pika==1.3.2
pytest==8.4.1